requirements.txt
beautifulsoup4==4.4.1
bleach==1.5.0
breadability==0.1.20
bs4==0.0.1
chardet==2.3.0
click==6.6
cycler==0.9.0
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.5
django-contrib-comments==1.7.3
docopt==0.6.2
facebook-sdk==0.4.0
filebrowser-safe==0.4.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
formencode==1.3.0
future==0.15.2
google-api-python-client==1.5.0
grappelli-safe==0.4.5
gunicorn==19.4.5
html5lib==0.9999999
httplib2==0.9.2
imdbpy==5.0
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
matplotlib==1.5.0
Mezzanine==4.2.2
nltk==3.2.1
numpy==1.6.2
oauth2client==2.0.1
oauthlib==2.0.0
pbr==1.9.1
Pillow==3.4.2
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pydispatcher==2.0.5
pyparsing==2.0.7
python-dateutil==2.4.2
python-telegram-bot==3.4
python-twitter==3.1
pytz==2016.7
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.7.0
rsa==3.4.2
simplejson==3.8.2
SQLAlchemy==1.0.12
sqlalchemy-migrate==0.10.0
sqlobject==3.0.0a2.dev20151224
sqlparse==0.1.19
sumy==0.4.1
tempita==0.5.3.dev0
tzlocal==1.3
uritemplate==0.6
virtualenv==15.0.1
Werkzeug==0.11.11
whitenoise==3.0
